refactor some database stuff
