+++ app/admin/views/boathouse.py
@@ -4,7 +4,8 @@ from flask_admin.contrib.sqla import tools
 
 from sqlalchemy.orm import Session
 
-from app.data import Boathouse, cache
+from app.data.models.boathouse import Boathouse
+from app.data.globals import cache
 from app.admin.base import ModelView
 from app.data.database import db
