+++ app/data/__init__.py
@@ -3,16 +3,7 @@
 The data module contains exactly what you'd expect: everything related to data
 processing, collection, and storage.
 """
-from .globals import cache
 from .database import db
-from .celery import celery_app
 
-# SqlAlchemy database models
-from .models.boathouse import Boathouse
-from .models.website_options import WebsiteOptions
-from .models.prediction import Prediction
-from .models.reach import Reach
-
-# External data calls
-from .processing import usgs
-from .processing import hobolink
+# Register to metadata.
+from .models import _all
