+++ app/data/models/_all.py
@@ -1,0 +1,7 @@
+# These imports are placed here to ensure that the SQLAlchemy models are always
+# registered to the db object's metadata.
+
+from .boathouse import Boathouse
+from .website_options import WebsiteOptions
+from .prediction import Prediction
+from .reach import Reach
