+++ app/data/queries/schema.sql
@@ -20,24 +20,11 @@ CREATE TABLE hobolink (
     air_temp        decimal
 );
 
-DROP TABLE IF EXISTS model_outputs;
-CREATE TABLE model_outputs (
-    reach           int,
-    time            timestamp,
-    log_odds        decimal,
-    probability     decimal,
-    safe            boolean
-);
-
 CREATE TABLE IF NOT EXISTS override_history (
     time            timestamp,
     boathouse_name  text,
     overridden      boolean,
     reason          text
 );
-
-
-DROP TABLE IF EXISTS boathouses;
-DROP TABLE IF EXISTS live_website_options;
 
 COMMIT;
